// backend/src/types/vehicle.types.ts
export interface CreateVehicleDto {
  make: string;
  model: string;
  year: number;
  category: string;
  number?: string;
  vin?: string;
  notes?: string;
}

export interface UpdateVehicleDto {
  make?: string;
  model?: string;
  year?: number;
  category?: string;
  number?: string;
  vin?: string;
  notes?: string;
}